+++ src/components/Admin/Admin.jsx
@@ -1,88 +1,20 @@
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
 import usePlaces from '../../hooks/usePlaces';
-import supabase from "../../components/api/supabaseClient";
+import supabase from "../api/supabaseClient";
+import {
+  Container,
+  Header,
+  Logo,
+  SearchBar,
+  EventList,
+  EventItem,
+  EventDetails,
+  ButtonGroup,
+  Button,
+} from './Admin.styled';
 
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  background-color: #f5f5f5;
-  height: 100vh;
-`;
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 10px 20px;
-  background-color: #fff;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-`;
-
-const Logo = styled.div`
-  font-size: 1.5em;
-  font-weight: bold;
-  color: #2d5f2e;
-  margin: 0 auto;
-`;
-
-const SearchBar = styled.input`
-  padding: 10px;
-  border-radius: 5px;
-  border: 1px solid #ddd;
-  width: 50%;
-  margin: 0 auto;
-`;
-
-const EventList = styled.div`
-  width: 100%;
-  max-width: 800px;
-  margin-top: 20px;
-`;
-
-const EventItem = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: #fff;
-  padding: 15px 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  margin-bottom: 15px;
-  cursor: pointer;
-`;
-
-const EventDetails = styled.div`
-  flex: 1;
-  margin-right: 20px;
-  font-size: 1.1em;
-  color: #333;
-`;
-
-const ButtonGroup = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const Button = styled.button`
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  background-color: #2d5f2e;
-  color: white;
-  cursor: pointer;
-  font-size: 0.9em;
-
-  &:hover {
-    background-color: #3a7741;
-  }
-`;
-
-const AdminPage = () => {
+const Admin = () => {
   const { data, error, isLoading } = usePlaces();
   const [places, setPlaces] = useState([]);
   const [searchTerm, setSearchTerm] = useState('');
@@ -164,4 +96,4 @@ const AdminPage = () => {
   );
 };
 
-export default AdminPage;
+export default Admin;
