--- src/page/Admin/Admin.styled.jsx
@@ -1,1 +1,0 @@
-import styled from "styled-components";
