+++ src/page/AdminPage.jsx
@@ -1,0 +1,10 @@
+import Admin from "../components/Admin/Admin";
+import React from 'react'
+
+export const AdminPage = () => {
+  return (
+    <>
+        <Admin />
+    </>
+  )
+};
