+++ src/page/AdminPostPage.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import AdminPost from '../components/Admin/AdminPost'
+
+export const AdminPostPage = () => {
+  return (
+    <>
+        <AdminPost />
+    </>
+  )
+};
