+++ src/routers/MainRouter.jsx
@@ -4,8 +4,8 @@ import SignUp from "../components/authentication/SignUp";
 import Login from "../components/authentication/Login";
 import DetailPage from "../page/DetailPage";
 import DefaultLayout from "../layouts/DefaultLayout";
-import AdminPostPage from "../page/Admin/AdminPostPage";
-import AdminPage from "../page/Admin/AdminPage";
+import AdminPostPage from "../components/Admin/AdminPost";
+import AdminPage from "../components/Admin/Admin";
 import Introduction from '../components/authentication/Introduction';
 import MyPage from '../page/MyPage';
 import { AdminRouters } from "./AdminRouters";
